Default to sha256 for RFC3161 timestamping of ClickOnce artifacts (#8801)
ClickOnce does not enforce SHA256 digest algo during RFC3161 timestamping of signed artifacts. For binaries, this means that depending on the version of signtool used, SHA1 is selected as the default digest algorithm.

All timestamp servers support SHA256 and there are some timestamp servers that don't support SHA1 anymore. The current default of SHA1 as the TS algo fails on TS servers that don't support it.

ClickOnce signing task has been updated to default to sha256 as the timestamping digest algo for both binary and manifest signing. For binaries, this is done by passing the /td switch to signtool.exe. For the XML manifests, this is done by passing the sha256 algo id to the CryptRetrieveTimeStamp API.