Containers: fix handling of ContainerBaseImage without a registry.
These images are meant to be pulled from Docker Hub.

The wrong hostname was used for the registry ('registry-1.docker.io' vs 'docker.io').

The parsing of names like 'ubuntu/dotnet-runtime' lead to 'ubuntu' being treated as the
registry.

This also enables using 'docker.io' as an alias to the Docker registry.

And adds the required 'library/' prefixing for short-names (like 'ubuntu' to 'library/ubuntu').