Reduce allocations in ConditionEvaluator.UpdateConditionedPropertiesTable
In order to parse the name of a property from an MSBuild expression, this method used a regular expression to parse a simple pattern, which results in significant amounts of allocations as identified by GCPauseWatson.

The actual pattern used is very simple and can be achieved regex.

This change extracts the parsing to a helper function and adds a unit test for it. I validated that the test passed with the prior implementation, then converted it to avoid regex allocations altogether.