Reduce allocations during SolutionFile parsing
This change converts much of the string processing to work with `ReadOnlySpan<char>` to help reduce allocations when parsing a solution file.

Almost all use of `string.Split` was removed, as this is a significant contibutor to GC pressure in this code.

Some portions of the parsing logic were pulled out to internal helper functions, and unit tests added. These tests were run on the prior implementation to ensure equivalent behaviour after these changes.

API documentation was outdated in a few places, so brought up to date.

Calls to `ProjectFileErrorUtilities.VerifyThrowInvalidProjectFile` require an instance of class `BuildEventFileInfo` to be passed. When no error condition is met, this object is unused. Several callers of this method were converted to performing their own check on the condition and calling `ProjectFileErrorUtilities.ThrowInvalidProjectFile` instead when a failure exists.

Other minor formatting changes, mostly guided by IDE diagnostics.