Add tests for GetSolutionFileAndVisualStudioMajorVersions
I spotted a bug in this method. This change makes the method more easily unit testable (without having to write to disk) and adds a failing test.

The bug occurs when certain solution header lines are indented. I will fix it in the next commit.

The large diff here is mostly just undenting the implementation after moving the try/finally to the new calling method.