Add support for starting the application through CMD (instead of ENTRYPOINT).
The new ContainerAppCommand and ContainerAppCommandArgs properties are used to hold
the command line to start the application.

The ContainerAppCommandInstruction controls how the command line is emitted.

The default behavior is to emit an ENTRYPOINT when the base image has no ENTRYPOINT,
and a CMD when the base image has an ENTRYPOINT (to use the base image ENTRYPOINT).

This also adds a ContainerCmd property, which can be used to directly set the image CMD.