add review comment to sb files (#8824)
Context
Contributes to dotnet/source-build#3435

Adds comments to source-build files asking for the inclusion of the source-build team in PRs that alter SourceBuild* files. Non-reviewed changes could potentially cause issues down the line, be it in the downstream repos or the product build (as has happened in the past, see dotnet/source-build#3435 (comment))

Changes Made
added comments to source-build files asking for the inclusion of the source-build team in PRs that alter SourceBuild* file.