Avoid boxing when enumerating project xml children
In a recent profile of a graph construction, it was observed that a
large amount of boxing was happening for
ProjectElementSiblingEnumerable. This change simplifies how xml children
are enumerated by adding an internal ChildrenEnumerable property which
directly exposes the ProjectElementSiblingEnumerable which should avoid
boxing, at least in some code paths (the public API makes it hard to
avoid everywhere...).

Additionally, a very common usage of enumerating children was to do
Children.OfType<T> and wrap it in a ReadOnlyCollection<T>, so I
introduced a GetChildrenOfType (and GetChildrenReversedOfType) method
which exposes an ICollection<T> which does the same thing but without
the boxing of ProjectElementSiblingEnumerable and without the OfType
class. It's just 1 collection allocation.