Remove references to BinaryFormatter from comments (#8851)
Related to #8827

Context
Some of the comments mentioning BinaryFormatter are misleading. Most of the classes don't need to be serializable. The only exception is PortableLibraryFile which is serialized cross-domain on .NET Framework.

Changes Made
Removed [Serializable] and the comment about BinaryFormatter-based calls from several classes. Left it on PortableLibraryFile, ifdef'ed to .NET Framework builds only.

Testing
Experimental insertion confirmed that PortableLibraryFile is indeed marshaled cross-domain in some scenarios. Confirmed that it is the only such type by code inspection.