Allow using latest illink analyzer on netstandard projects (#33158)
Before net8.0, the illink analyzer shipped as a separate package
from ILLink.Tasks. https://github.com/dotnet/sdk/pull/32045 added
logic to reference the separate package for tfms less than
net8.0.

This change tweaks the logic to be based on the package version
resolved from KnownILLinkPack, rather than the tfm. This makes it
possible to define KnownILLinkPack to pull in the latest
analyzer, as a workaround when targeting unsupported TFMs. The
scenario is still considered unsupported, but this at least makes
it possible to work around.

A new testcase shows the steps necessary to use the latest illink
analyzer on a netstandard2.x project.