Removing sln level turn off of setplatform feature
Currently we turn off dynamic platform resolution for a whole solution
if a single project in the solution is assigned a configuration. This is
problematic as some projects are outside of the scope of the solution
but still have certain targets that run on them that are architecture
specific. These projects will build as the wrong architecture because no
configuration is defined and no platform negotiation takes place.

I removed the conditional that turns platform negotiation off on a sln
level. The logic to turn this off on a project level is already in place
through checking is a projectreference has setplatform appended to it.
This will make sure no projects with configurations defined will be
negotiated for as MSbuild adds setplatform metadata to projectreferences
with configurations.