Avoid updating lines when colorized
The root cause of #8781 is that the optimize-output-by-skipping-
identical-characters logic is not aware of escape sequences, so it goes
horribly awry when the replacing line differs after the first escape
sequence.

Instead, apply the previous logic only if there are no escape sequences;
if there are, replace the entire line.

If this proves to cause too much flicker or perform slowly, we could
implement escape-sequence-aware truncation, length, and comparison, but
we can wait for feedback to that effect.