Fixed #8963
--
Updated the SolutionFile.FullPath property to have a public getter and an internal setter.  Since, now the property is part of a publicly-exposed API, reworked the XML documentation for the property as well.

Updated the SolutionFile.SolutionFileDirectory property to have a public getter and an internal setter.  I also added XML documentation configuration for it.