Fix "Registry is not supported on this platform" with mono
Running `dotnet --info` with mono fails with:

[ERROR] FATAL UNHANDLED EXCEPTION: System.TypeInitializationException: The type initializer for 'Microsoft.Win32.Registry' threw an exception.
 ---> System.PlatformNotSupportedException: Registry is not supported on this platform.
   at Microsoft.Win32.RegistryKey.OpenBaseKeyCore(RegistryHive hKey, RegistryView view)
   at Microsoft.Win32.RegistryKey.OpenBaseKey(RegistryHive hKey, RegistryView view)
   at Microsoft.Win32.Registry..cctor()
   --- End of inner exception stack trace ---
   at Microsoft.DotNet.Cli.Utils.UILanguageOverride.Setup()
   at Microsoft.DotNet.Cli.Program.InitializeProcess()
   at Microsoft.DotNet.Cli.Program.Main(String[] args)

The culprit is the CurrentPlatformIsWindowsAndOfficiallySupportsUTF8Encoding()
method: even though it wraps the usage of the Microsoft.Win32.Registry
class in an "if" statement, the spec allows loading this class when
JITing the method where it's used, leading to the failure above.

Fix by moving the Microsoft.Win32.Registry usage into a separate
method. Add Debug.Assert() in order to pacify the checker.

Fixes: b6ca751324b8 ("[wip] restore encoding and only use on win that support utf 8 oficially")