improve TaskItem hashcode (#8988)
Tiny improvement to the hash code calculation perf - ItemSpec is just EscapingUtilities.UnescapeAll(_includeEscaped), which is a simple deterministic transform so we can just use the string.

Added a comment to explain why we can't easily hash in something about the metadata. This means that all items with the same item spec continue to have the same hash code, which can cause degenerate performance.