Right justify output
Move cursor right (it stops at end of line), then back to justify the
output. This makes the localization less flexible since the formatting +
justification is handled in code, but makes it a lot easier to
understand. In the future reordering in translations would need to be
handled fairly carefully.

Avoid a bunch of overwrite problems by simplifiying the
should-we-do-a-fancy-overwrite decision to "is just time tick" or "just
overwrite it all".