Fix single-file static web assets tests (#34186)
Fixes test bugs in two tests.  These tests were calling
`Execute($"/p:PublishSingleFile=true
/p:RuntimeIdentifier={RuntimeInformation.RuntimeIdentifier}")`,
which was setting the value of `PublishSingleFile` to `true
/p:RuntimeIdentifier=<RID>`. The two MSBuild properties should
have been passed as separate arguments to `Execute`.

This was resulting in the tests not actually publishing
single-file, and not setting the RID. Fixing this required
updating the tests to expect RID-specific static web assets. The
expected manifests were then updated by running the
`update-test-baselines.ps1` script.