Add separate warnings for settings that depend on illink pack (#34077)
This splits the error added in
75ce41dce3074d15298f93e97ba95af12658dcbf into separate
warnings/errors specific to the various options that depend on
the ILLink pack. These warnings/errors are produced when those
options are set for an unsupported target framework.

The PublishTrimmed error text reuses the message that used to be
produced as a warning when targeting < netcoreapp3.0. This error
is produced only when PublishAot is not set (there's a more
specific error if it is).

Previously, we would produce errors when attempting to use the
analyzers on unsupported TFMs. This has been changed to a warning
in the case of a library (for example IsTrimmable, rather than
PublishTrimmed). See comments in ProcessFrameworkReferences for
details.

This also fixes an issue where ProcessFrameworkReferences was not
running for netcoreapp2.0 because there were no
FrameworkReferences defined for that TFM. This adds an extra
condition to let ProcessFrameworkReferences run if any settings
depend on the ILLink pack, so that it can produce proper
warnings/errors.