Add back _RequiresILLinkPack override (#34429)
bdce224bba472ccad4bade07b757ac5275040c0e removed the ability to
download the ILLink pack by setting `_RequiresILLinkPack`. This
is needed for macios scenarios where the ILLink pack should be
restored, but it's not known ahead-of-time whether trimming will
be enabled. See
https://github.com/xamarin/xamarin-macios/pull/17227 for more
context.

This adds back the ability to set `_RequiresILLinkPack` to
download the ILLink pack even if none of the other relevant
settings are enabled.