Containers: prefer starting the app using 'dotnet <assembly.dll>'.
The default apphost, which comes with the SDK, may be incompatible
with the container base image.

For example: the apphost of a musl-based Linux SDK is not compatible
with the default (glibc Debian) based Microsoft base images.

And, a source-build glibc based apphost may also be incompatible
with the Microsoft glibc based images due to glibc symbol versioning.