Expose Binlog reader API needed for sensitive data redacting (#9132)
Contributes to #9089

Context
The simpliest way of redacting specific strings from binlogs is by replaying it via reader, while altering strings and immediately writing to new binlog via writer. To allow this scenario - the string manipulation during reading is exposed.
Based on team decision - the actual leveraging of the exposed functionality will currently live outside of this repo

Changes Made
Added interfaces with events
Added EventArgs allowing manipulation of the read strings (return value or ref argument of an event is an antipattern. Alternative we could use a Span<char> as an event arg - allowing in-place changes - but that would mean we'd need to create the span [which involves copying of the internal string buffer], for each read string, even those we wouldn't need to alter)
Made reader code nullable compatible