Add packages to Version.Details for source-build (#9172)
This fixes a portion of the assemblies described in https://github.com/dotnet/source-build/issues/3599 that are showing up as reference assemblies in the source-built SDK. The reason they're showing up as ref assemblies is the same reason that's described in https://github.com/dotnet/msbuild/pull/9158, which is a related issue.

It's not known whether the existence of these ref assemblies causes a functional issue as it does with https://github.com/dotnet/msbuild/pull/9158. But we do know that the source-built 7.0 SDK doesn't define these as ref assemblies but rather implementation assemblies. So to maintain parity with 7.0 and avoid potential risk, it's best to ensure these are represented as implementation assemblies in the output.

These are the assemblies that are fixed by these changes:

* System.CodeDom
* System.Resources.Extensions
* System.Security.Cryptography.Xml