Check console mode in -get* message output (#9094)
Take into account console mode when writing error messages as part of a user asking for a particular property/item/target result. If we can't verify that the console is VT-compatible, we fall back to just printing things out with their default coloration. (Though not perfect, I think this is an unimportant enough scenario, that I don't think it's worth worrying about.)

This also includes code cleanup.

Fixes #9079