ProjectCache: Break away from UI thread for intialization to avoid VS hangs (#9236)
This fixes a hang reported by some users which appears to impact some specific scenarios.

For some legacy scenarios in VS, it appears that design-time builds are not properly detected by the project cache and so trigger the scenario. This runs on the UI thread, which itself is bad, but then in the `InitializePluginsForVsScenario` method, some tasks are used which also execute on the UI thread and the specific plugin the customer is using makes many async calls which ends up hanging the UI thread.

Originally, I had wrongly assumed that the `Parallel.ForEach` would cause the code in the body to run in a separate thread, breaking away from the UI thread. However, that's a bad assumption as with a small number of items (1 in this case) it just runs synchronously.

This just slaps it in a `Task.Run`, which is probably inelegant, but hey it works. I'm open to suggestions, but I do think it's important to "break away" entirely since we would not want VS's UI thread to be beholden to plugin (external) code which may not `ConfigureAwait` properly, as is the case here :)

I applied the same change to `InitializePluginsForGraph` proactively since it does basically the same thing, although the VS scenario doesn't use it today, but might as well future-proof it.