Setting ReferenceOutputAssembly to falso on references to esproj (#9201)
Setting ReferenceOutputAssembly to falso on references to esproj
Esprojs may be referenced by ASP.Net projects (or others). When this happens,
the .net project will have to not reference output assembly on the esproj,
because there is no output assembly. This targets provides that behavior.
There was an argument for putting this in this file instead of in the sdk.
The discussion can be found here: https://github.com/dotnet/sdk/pull/34983