Translate FileAccessData in TaskHostTaskComplete
TaskHostTaskComplete packets can be sent between nodes that differ in
FEATURE_REPORTFILEACCESS, causing confusion when the sending side (for
example a net35 taskhost) does not send the FileAccessData field, but
the receiving side (net48) expects it.

Unify this by explicitly sending a bool `false` (no data) in the
!FEATURE_REPORTFILEACCESS case, so there's no difference on the
receiving end between "no data" and "couldn't be any data".