Use AssemblyLoadContext Name instead of AppDomain (#9353)
When MSBuild was updated to isolate tasks into a custom `AssemblyLoadContext`, `AssemblyLoadsTracker` was not updated to be aware of this change. This means that the log messages always say `AppDomain: [Default]` and don't show any information about which `AssemblyLoadContext` was used to load the assembly.

When `FEATURE_ASSEMBLYLOADCONTEXT` is defined:
- `AssemblyLoadsTracker` now populates `AssemblyLoadBuildEventArgs` with the name of the `AssemblyLoadContext` used to load the assembly . The `appDomainDescriptor` parameter is re-used to do this.
- A new string resource has been added to update the text of `AssemblyLoadBuildEventArgs.Message` to use label the `AppDomainDescriptor` value as "AssemblyLoadContext" instead of "AppDomain".

Fixes #9348.