Exec: use C.UTF-8 instead of en_US.UTF-8 to set the UTF-8 encoding.
The en_US locale can't be used on systems where it is not installed.
This is common in container images.

On such systems, setting the locale to en_US.UTF-8 causes unexpected
warnings to be written to standard error.

When Exec.LogStandardErrorAsError is set, these warnings cause
the Task to fail due to logging errors.

This changes to use the 'Computer English' C.UTF-8 locale, which is
always available.