Log allowed environment vars case-insensitively
I was recently investigating a problem where the issue was that
`MSBuildSdksPath` was set incorrectly, so an SDK was being resolved
incorrectly. I had initially ruled that out, because I knew that we
should be logging environment variables that start with `MSBUILD`.

We were, but case-sensitively. Since the canonical case for some of our
config environment variables is not all caps, we should change that.