Update Microsoft.BuildXL.Processes (#9476)
Update Microsoft.BuildXL.Processes

The primary motivation behind this update is to bring in a bug fix related to the BuildXL.Process assembly depending on BuildXL.Tracing, which is not available in the package.

This change also adjusts a breaking change to ChildProcessesToBreakawayFromSandbox, as well as added some extra logic associated with the reason behind the breaking change which supports dotnet scenarios better. This is just future-proofing though since reporting file accesses via detours is currently only available in MSBuild.exe.