Fixed /restore and /graph conjunction error on exit code (#9461)
Fixes #9443

Context
When the /graph and /restore are used within the same command, the build exit code will always be 0. This is because of the variable used to define success of the restore action overrides the success of the graph build.

Testing
Made sure existing tests passed and added a unit test for this case, an some manual testing.