Pass submissionId to the BuildEventContext used by SolutionProjectGenerator (#9470)
* Pass submissionId to the BuildEventContext used by SolutionProjectGenerator

When evaluating the generated .sln.metaproj, we do pass a valid submissionId, but then we don't use it when creating the BuildEventContext used by the ProjectInstance to log evaluation events. So evaluation logging events end up having SubmissionId == -1, and thus the loggers registered with the MuxLogger in the IDE do not get the evaluation events.

Fixes #9469

* Ensure that BuildEventContext passed to LoadSolutionForBuild has a valid SubmissionId

This way Metaproject generated and other events get properly routed to MuxLogger and other loggers that care about the SubmissionId.