Logger emits message on invalid property name (#9570)
Fixes #9475

Context
When a property name is an invalid input, we log the error and shutdown the loggers when terminating the process. However, we did not flush the logs of any messages being processed before shutting down, making we miss this specific error in the process.

Changes Made
Added a wait for loggers to finish doing their work before shutting down.

Testing
Added a test to make sure that the logger emits the message before shutdown.