Update Traits when environment has been changed (#9655)
Fixes #9501

Context
When the new environment is sent to msbuild node or msbuild server node, the environment is set up but the values in the Traits class are not updated. This leads to using a different configuration on the server or MSBuild node than on the main MSBuild node when the server or msbuild node was preserved from the previous builds.

Changes Made
Together with setting the new environment, re-create a Traits class instance to update the corresponding values.
I bit of refactoring: there was a pattern in the code "unset current environment and set a new one", which is repeated in many places.

Testing
Manual tests, unit tests, exp VS insertion

Notes
This change is rather a work-around for the bigger problem of handling the environment properly in MSBuild nodes: the configuration sometimes is taken from the Traits and sometimes directly from environment variable. Also, as in this issue, sometimes code sets new environment variables but does not update the values in Traits (and sometimes it updates). We might consider fixing how we deal with configuration.