Checking for VS pathing when import not found (#9625)
Fixes #8967

Context
Users sometimes use different distributions to build projects, however some files available in the VS distribution are not available with dotnet build. It used to throw a "File not found" error, but now it is switched to "this imported project is in a VS distribution".

An example message is:  error MSB4278: The imported project $(MSBuildE xtensionsPath)\Microsoft\VisualStudio\v$(VisualStudioVersion)\DSLTools\Microsoft.DslTools.settings.targets is from the Visual Studio distribution of MSBuild. Build with MSBuild.exe in a Developer Command Prompt.

Changes Made
Before throwing a invalidProject error for file not found, check the path of the file if it is within Visual Studio. If it is, throw new error instead.