Support environments with long variable values (#9794)
Fixes [AB#1527557](https://devdiv.visualstudio.com/0bdbc590-a062-4c3f-b0f6-9383f67865ee/_workitems/edit/1527557)

### Context

The .NET Framework implementation of `Environment.SetEnvironmentVariable` has a hard-coded check for the 32,767 character limit and exceeding it fails the call with `ArgumentException`. The limit does not exist anymore on modern Windows and we've started seeing users run MSBuild in environments with very long variables. This is problematic because MSBuild contains logic to restore the environment, so even if the particular variable is not really used, we may end up trying to set it to an excessively long value.

### Changes Made

- Made `CommunicationUtilities.SetEnvironment` use a P/Invoke to `kernel32!SetEnvironmentVariable` on .NET Framework, to work around the problematic BCL implementation.
- Optimized `CommunicationUtilities.SetEnvironment` to call `SetEnvironmentVariable` only if it doesn't already have the desired value.
- Made `CommunicationUtilities.SetEnvironment` calls to restore the environment _best-effort_ in places where the process is anyways shutting done.
- Changed `CommunicationUtilities.GetEnvironmentVariables` to use the custom implementation only on .NET Framework where it provides the perf benefit. On .NET Core it is actually slower than the BCL implementation.

### Testing

Added a new unit test class and also pulled one existing test to it because it was unnecessarily running twice in its previous location.