Maybe permit failed SDK load
If an SDK is a workload SDK, and we don't have the workload SDK resolver from the SDK, as, for instance, if we have an xcopy'd MSBuild, then we should still be able to build simple projects that don't require workloads. Right now, we just fail. If we get into a failure state, this tries to figure out if we actually need workloads, erring on the side of failing if we aren't sure. If we decide we don't need workloads, however, we suppress the error for failing to load a workload SDK.

Fixes #7988