Notify MSBuild TerminalLogger of project cache plugin starting (#10027)
Adds information per-project to determine whether the project is classified as a cache plugin project, which occurs when the cache plugin target is called (prior to `GetTargetPath` or `Build`). For cache plugin projects that are built, the scenario remains the same (the output path is shown). For cache plugin projects that result in a hit, indicate the output path as well to indicate build progression.