Expose ToolTask exit, stdout, and stderr handlers (#10148)
### Context

The `StartToolProcess` method added in #9649 does not allow subclasses to create and return equivalent processes because it does not have access to the exit, stdout, and stderr event handlers subscribed in the incoming `Process` object. In C# it is not possible to get the invocation list of an event outside of its declaring class.

### Changes Made

Made `ReceiveExitNotification`, `ReceiveStandardErrorData`, and `ReceiveStandardOutputData` protected to expose them to subclasses.

### Testing

A new unit test exercising the scenario.